Add telemetry for gremlin queries
